refactor: reorganize order status and parser files, update imports accordingly
